ECM-109: A few renamings
